Feature Extractor now saves to a property file with a given directory. Creates a file if it does not exist.
