--- src/main/java/FeatureExtraction/ExtractionAlgorithm.java
@@ -1,7 +1,0 @@
-package featureextraction;
-
-import java.awt.image.BufferedImage;
-
-public interface ExtractionAlgorithm {
-    Double[] extractFeatures(BufferedImage bufferedImage);
-}
