+++ src/main/java/Main.java
@@ -1,9 +1,13 @@
+import tletters.featureextraction.FeaturesExtractor;
+
+import java.io.IOException;
+
 /**
  * Created by mariuszbeltowski on 23/04/16.
  */
 
 public class Main {
-    public static void main(String[] args) {
-        System.out.println("Hello World!");
+    public static void main(String[] args) throws IOException {
+        FeaturesExtractor featuresExtractor = new FeaturesExtractor("xxxxxxxxxxxxx");
     }
 }
