+++ src/main/java/tletters/featureextraction/ExtractionAlgorithm.java
@@ -1,0 +1,7 @@
+package tletters.featureextraction;
+
+import java.awt.image.BufferedImage;
+
+public interface ExtractionAlgorithm {
+    double[] extractFeatures(BufferedImage bufferedImage);
+}
